+++ src/yw2nwlib/__init__.py
@@ -5,7 +5,9 @@ Modules:
 handles -- Provide a class to manage novelWriter handles.
 nw_configuration -- Provide a configuration class for reading and writing INI files.
 nw_converter -- Provide an novelWriter converter class for yWriter.
-nw_item -- Provide a class for novelWriter items.
+nw_item -- Provide a generic strategy class for novelWriter items.
+nw_item_v1_3 -- Provide a strategy class for novelWriter items (file format version 1.3).
+nw_item_v1_4 -- Provide a strategy class for novelWriter items (file format version 1.4).
 nwx_file -- Provide a class for novelWriter project file representation.
 nwd_file -- Provide a generic class for novelWriter item file representation.
 nwd_novel_file -- Provide a class for novelWriter novel file representation.